+++ CHANGELOG.md
@@ -1,7 +1,22 @@
-## v10.3.3 (unreleased)
+## v10.3.3 (September 13th 2026)
 ### Implementation changes and bug fixes
 - [PR #2770](https://github.com/rqlite/rqlite/pull/2770): Upgrade gRPC to 1.83.2.
-- [PR #2771](https://github.com/rqlite/rqlite/pull/2771): Refactor busy timeout methods. Thanks @ zxysilent
+- [PR #2771](https://github.com/rqlite/rqlite/pull/2771): Refactor busy timeout methods. Thanks @zxysilent
+- [PR #2776](https://github.com/rqlite/rqlite/pull/2776): Fix edge-case bugs detected by GPT-6 Astra.
+- [PR #2781](https://github.com/rqlite/rqlite/pull/2781): Switch to `quay.io/minio/minio:latest` for testing purposes.
+- [PR #2778](https://github.com/rqlite/rqlite/pull/2778): Fix removing a node with different ID, but same Raft address.
+- [PR #2782](https://github.com/rqlite/rqlite/pull/2782): Use correct Foreign Key setting when recovering node.
+- [PR #2779](https://github.com/rqlite/rqlite/pull/2779): Only dump the indexes and triggers of the selected tables. Fixes issue [#2777](https://github.com/rqlite/rqlite/issues/2777). Thanks @geiltonxavier
+- [PR #2783](https://github.com/rqlite/rqlite/pull/2783): Detect DB changes during Request(RETURNING).
+- [PR #2784](https://github.com/rqlite/rqlite/pull/2784): Snapshot `Open()` first checks for failed reap.
+- [PR #2785](https://github.com/rqlite/rqlite/pull/2785): Sync snapshot plan directory.
+- [PR #2786](https://github.com/rqlite/rqlite/pull/2786): Propagate WAL stat errors.
+- [PR #2787](https://github.com/rqlite/rqlite/pull/2787): Clean up cancelled snapshots.
+- [PR #2788](https://github.com/rqlite/rqlite/pull/2788): Reject incomplete snapshot headers.
+- [PR #2789](https://github.com/rqlite/rqlite/pull/2789): Validate restore database header.
+- [PR #2790](https://github.com/rqlite/rqlite/pull/2790): Only reap a sole fullsnapshot if it has zero WALs.
+- [PR #2790](https://github.com/rqlite/rqlite/pull/2793): Keep database dumps consistent during writes to the database.
+- [PR #2792](https://github.com/rqlite/rqlite/pull/2792): Roll back DB Requests on statement-prepare errors.
 
 ## v10.3.2 (September 7th 2026)
 ### Implementation changes and bug fixes